Minor code cleanup for AlignmentHelperLinearSpace.java and AlignmentHelper.java.
Simple change to PairwiseAligner.java to support the no-new-internal-gaps feature.
The change is a simple one: make internal gaps impossibly expensive, so they
won't happen. I've done this by setting the internal gap extension penalty to
Integer.MAX_VALUE/10  (implausibly high cost, but low enough that there's no risk
of two large numbers being added together to cause an overflow of the capacity
of an int).

In this implementation, a slight preference will be given to alignments that are
flush on one end, as this will reduce the number of terminal gap opens to one
instead of two. This can be easily overcome by changing the terminal gap open
to 0.  I've chosen not to make this the default behavior, as it will slightly
change terminal gap cost dynamics, and I don't know if that's for the better
or worse.